fix: convert all test files to use INTEGER cents instead of DECIMAL money values
- Updated all test schema definitions from DECIMAL(10,2) to INTEGER for money columns
- Converted all .00 decimal values to integer cents (multiplied by 100)
- Updated all test assertions to expect integer cent values
- Affected test files: analytics.test.js, dailyLedger.test.js, dailySummary.test.js,
  dashboard.test.js, directorWithdrawal.test.js, expense.test.js, importExport.test.js,
  income.test.js, incomeCategory.test.js, report.test.js, student.test.js,
  studentCharge.test.js, transaction.test.js

This completes the money-to-cents migration for test files (RC8).

+++ backend/src/__tests__/analytics.test.js
@@ -89,7 +89,7 @@ describe('Analytics Model', () => {
       CREATE TABLE IF NOT EXISTS transactions (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         receipt_number TEXT UNIQUE NOT NULL,
-        amount DECIMAL(10, 2) NOT NULL,
+        amount INTEGER NOT NULL,
         transaction_type TEXT NOT NULL,
         description TEXT,
         related_id INTEGER,
@@ -176,7 +176,7 @@ describe('Analytics Model', () => {
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         name TEXT NOT NULL,
         description TEXT,
-        amount DECIMAL(10, 2) NOT NULL,
+        amount INTEGER NOT NULL,
         charge_type TEXT NOT NULL,
         frequency TEXT NOT NULL,
         is_recurring BOOLEAN DEFAULT 0,
